Implement foolsgold shared nothing rack app with nemesis::handler
