+++ foolsgold/Cargo.toml
@@ -22,9 +22,6 @@ rocket = "0.4.0"
 [dependencies.mruby]
 path = "../mruby"
 
-[dependencies.mruby-gems]
-path = "../mruby-gems"
-
 [dependencies.nemesis]
 path = "../nemesis"
